Two functions added
bool Org_fitness(Organism *org, double age, double food_gain, int
wall_hit)   --> updates the fitness of the org and return true if the
org is a winner

double alt_penalize(double giver_old_food_level, double food_granted,
double rec_old_food_level, double r)
--> This function compute the penalty for those irrational altruistic
agents for not satisfying the hamilton rule.  if this rule is satisfied
it does not penalize the agent and returns 0 otherwise a minus value is
return which the magnitude indicates how that altruistic behavior was
Improper (for example the absoultue value is the largest when the agent
sacrifice itself).